+++ src/main/scala/colinh/model/Comment.scala
@@ -3,11 +3,11 @@ package colinh.model
 import org.squeryl.PrimitiveTypeMode._
 
 class Comment(
-    var id:        Long,
-    var articleId: Long,
+    var id:        Int,
+    var articleId: Int,
     var body:      String,
     var createdAt: DateType,
     var updatedAt: DateType,
-    var userId:    Long) {
+    var userId:    Int) {
   def this() = this(0, 0, "", null, null, 0)
 }
